refactor: extract participants merge helpers into conversion_participants_merge.py
